requirements.txt
# VMS CLI dependencies
click>=8.0.0
colorama>=0.4.4
tabulate>=0.9.0
pyyaml>=6.0
google-cloud-compute>=1.8.0
google-cloud-resource-manager>=1.10.0
google-auth>=2.0.0
requests>=2.28.0